+++ lib/src/ui/models.dart
@@ -1,0 +1,9 @@
+import '../utils/annotations.dart';
+
+@experimental
+class IntentResponse {
+  final String status;
+  final String clientSecret;
+
+  IntentResponse(this.status, this.clientSecret);
+}