+++ lib/src/utils/annotations.dart
@@ -1,0 +1,1 @@
+const Deprecated experimental = Deprecated("experimental");