+++ lib/stripe_sdk_ui.dart
@@ -4,6 +4,8 @@ export './src/card_number_formatter.dart';
 export './src/card_utils.dart';
 export './src/model/card.dart';
 export './src/ui/add_payment_method.dart';
+export './src/ui/checkout_screen.dart';
+export './src/ui/models.dart';
 export './src/utils/card_utils.dart';
 export './src/widgets/card_cvc_form_field.dart';
 export './src/widgets/card_expiry_form_field.dart';